PhoneBook - Reorganisation
Project reorganisation, files and some fixes in code.
