--- Class Headers/Password.h
@@ -1,17 +1,0 @@
-#ifndef PASSWORD_H_INCLUDED
-#define PASSWORD_H_INCLUDED
-
-#include "IPassword.h"
-
-class Password: public IPassword
-{
-    char *c;
-public:
-    Password();
-    ~Password();
-    void AddPassword();
-    int ReturnPass();
-};
-
-
-#endif // PASSWORD_H_INCLUDED
