+++ Interfaces/IContact.h
@@ -4,9 +4,7 @@
 class IContact
 {
 public:
-   // void AddId() = 0;
-    virtual void Print() = 0;
-    virtual char* Get()  = 0;
+	virtual char* Get()  = 0;
 };
 
 #endif // ICONTACT_H_INCLUDED
