+++ Interfaces/IPerson.h
@@ -6,7 +6,7 @@ class IPerson
 {
 public:
     virtual void AddPersoana() = 0;
-    virtual  void Remove()     = 0;
+    virtual void Remove()      = 0;
 };
 
 #endif // IPERSON_H_INCLUDED
