+++ Sources/Console.cpp
@@ -1,0 +1,13 @@
+#include "../Headers/Console.h"
+#include "../Headers/Person.h"
+
+namespace Console
+{
+	void gotoxy(int x, int y)
+	{
+		COORD coord;
+		coord.X = x;
+		coord.Y = y;
+		SetConsoleCursorPosition(GetStdHandle(STD_OUTPUT_HANDLE), coord);
+	}
+}
