+++ Sources/PhoneBook.cpp
@@ -1,0 +1,13 @@
+#include <iostream>
+#include "../Headers/Object.h"
+#include "../Headers/Admin.h"
+#include <fstream>
+
+using namespace std;
+
+int main()
+{
+	ILevel *level = new Level();
+	level->PrintItems();
+	return 0;
+}
