+++ menu.cbp
@@ -1,0 +1,72 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
+<CodeBlocks_project_file>
+	<FileVersion major="1" minor="6" />
+	<Project>
+		<Option title="PhoneBook" />
+		<Option pch_mode="2" />
+		<Option compiler="gcc" />
+		<Build>
+			<Target title="Debug">
+				<Option output="bin/Debug/menu" prefix_auto="1" extension_auto="1" />
+				<Option object_output="obj/Debug/" />
+				<Option type="1" />
+				<Option compiler="gcc" />
+				<Compiler>
+					<Add option="-g" />
+				</Compiler>
+			</Target>
+			<Target title="Release">
+				<Option output="bin/Release/menu" prefix_auto="1" extension_auto="1" />
+				<Option object_output="obj/Release/" />
+				<Option type="1" />
+				<Option compiler="gcc" />
+				<Compiler>
+					<Add option="-O2" />
+				</Compiler>
+				<Linker>
+					<Add option="-s" />
+				</Linker>
+			</Target>
+		</Build>
+		<Compiler>
+			<Add option="-Wall" />
+			<Add option="-fexceptions" />
+		</Compiler>
+		<Unit filename="Headers/Admin.h" />
+		<Unit filename="Headers/Console.h" />
+		<Unit filename="Headers/Email.h" />
+		<Unit filename="Headers/Id.h" />
+		<Unit filename="Headers/Level.h" />
+		<Unit filename="Headers/Name.h" />
+		<Unit filename="Headers/Number.h" />
+		<Unit filename="Headers/Object.h" />
+		<Unit filename="Headers/Password.h" />
+		<Unit filename="Headers/Person.h" />
+		<Unit filename="Headers/User.h" />
+		<Unit filename="Headers/Username.h" />
+		<Unit filename="Interfaces/IContact.h" />
+		<Unit filename="Interfaces/ILevel.h" />
+		<Unit filename="Interfaces/IPassword.h" />
+		<Unit filename="Interfaces/IPerson.h" />
+		<Unit filename="Interfaces/IUsername.h" />
+		<Unit filename="Sources/Admin.cpp" />
+		<Unit filename="Sources/Console.cpp" />
+		<Unit filename="Sources/Email.cpp" />
+		<Unit filename="Sources/Id.cpp" />
+		<Unit filename="Sources/Level.cpp" />
+		<Unit filename="Sources/Name.cpp" />
+		<Unit filename="Sources/Number.cpp" />
+		<Unit filename="Sources/Object.cpp" />
+		<Unit filename="Sources/Password.cpp" />
+		<Unit filename="Sources/Person.cpp" />
+		<Unit filename="Sources/PhoneBook.cpp" />
+		<Unit filename="Sources/User.cpp" />
+		<Unit filename="Sources/Username.cpp" />
+		<Extensions>
+			<code_completion />
+			<envvars />
+			<debugger />
+			<lib_finder disable_auto="1" />
+		</Extensions>
+	</Project>
+</CodeBlocks_project_file>
